Addresed comments
