+++ .env
@@ -1,0 +1,6 @@
+DB_USER = 'root'
+DB_PASSWORD = 'mysqleight'
+DB_HOST = 'localhost'
+DB_NAME = 'Test_schema'
+SERVER_PORT = 8080
+DB_PORT = 3306
